XEH - Move compats to individual addons
