+++ addons/ee/addon.toml
@@ -1,0 +1,2 @@
+[preprocess]
+enabled = false
